Stability improvements; dependency verification & bug fixes 	modified:   Sonic-06-Toolkit/Properties/AssemblyInfo.cs 	modified:   Sonic-06-Toolkit/Properties/Resources.Designer.cs 	modified:   Sonic-06-Toolkit/Properties/Resources.resx 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitEnvironmentX.cs 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitProgram.cs 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitSessionLog.Designer.cs 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitSessionLog.cs 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitText.cs 	modified:   Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitTools.cs 	modified:   Sonic-06-Toolkit/src/Tools/BINA/CollisionGenerator.cs 	modified:   Sonic-06-Toolkit/src/Tools/BINA/PlacementConverter.cs 	modified:   Sonic-06-Toolkit/src/Tools/BINA/TextEncoding.cs 	modified:   Sonic-06-Toolkit/src/Tools/DirectDraw/TextureConverter.cs 	modified:   Sonic-06-Toolkit/src/Tools/Lua/LuaCompilation.cs 	modified:   Sonic-06-Toolkit/src/Tools/SegaNN/XNOTool.cs 	modified:   Sonic-06-Toolkit/src/Tools/SonicAudioTools/SonicSoundStudio.cs 	modified:   Sonic-06-Toolkit/src/Tools/XexTool/ExecutableModification.cs

+++ Sonic-06-Toolkit/src/Sonic-06-Toolkit/ToolkitProgram.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.IO;
+using Toolkit.Text;
+using Toolkit.Tools;
 using System.Diagnostics;
 using System.Windows.Forms;
 using System.Security.Principal;
@@ -95,9 +97,11 @@ namespace Toolkit.EnvironmentX
             if (!File.Exists($"{applicationData}\\Hyper_Development_Team\\Sonic '06 Toolkit\\Tools\\s06col.exe"))
                 File.WriteAllBytes($"{applicationData}\\Hyper_Development_Team\\Sonic '06 Toolkit\\Tools\\s06col.exe", Properties.Resources.s06col);
 
-            //if (!File.Exists($"{applicationData}\\Hyper_Development_Team\\Sonic '06 Toolkit\\Tools\\s06collision.exe"))
+            if (!File.Exists($"{applicationData}\\Hyper_Development_Team\\Sonic '06 Toolkit\\Tools\\s06collision.exe"))
                 File.WriteAllBytes($"{applicationData}\\Hyper_Development_Team\\Sonic '06 Toolkit\\Tools\\s06collision.exe", Properties.Resources.s06collision);
-            
+
+            Verification.VerifyDependencies(Paths.Tools);
+
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new Main(args, sessionID));
